requirements.txt
selene>=2.0.0rc9
pytest>=8.1.1
allure-pytest>=2.13.5
python-dotenv == 1.0.0
jsonschema>=4.22.0
requests>=2.32.2
